LayoutTests/inspector/canvas/recording-expected.txt

Test general cases of CanvasAgent recording calls.


== Running test suite: Canvas.recording
-- Running test case: Canvas.multipleRecording
Starting a recording of canvas 1...
PASS: Recording started of canvas 1
Starting a recording of canvas 2...
PASS: Recording started of canvas 2
Performing actions...
PASS: Actions performed.
Stopping the recording of canvas 1...
PASS: There should be a recording for canvas 1.
Stopping the recording of canvas 2...
PASS: There should be a recording for canvas 2.

-- Running test case: Canvas.startRecording.InvalidCanvasId
PASS: Should produce an error.
Error: No canvas for given identifier.

-- Running test case: Canvas.stopRecording.InvalidCanvasId
PASS: Should produce an error.
Error: No canvas for given identifier.
